Refactor homepage and header: modularize components, enhance responsiveness, and add mobile navigation
